refactor: InputSection컴포넌트 common 위치로 이동 #1480
- InputWithSubmitButton 으로 이름 변경
- placeholder props 추가
